+++ packages/desktop/src/rtl-tests/community.create.test.tsx
@@ -203,6 +203,7 @@ describe('User', () => {
         "Files/checkForMissingFiles",
         "Network/addInitializedCommunity",
         "Communities/clearInvitationCodes",
+        "Communities/sendCommunityMetadata",
         "Identity/saveOwnerCertToDb",
         "PublicChannels/createGeneralChannel",
         "PublicChannels/channelsReplicated",