+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Replaced greying out inputs with splash screen on joining/creating screens.
 
+* Fixes empty space between chat's input and a soft keyboard on iOS devices.
+
 * Changed registration process - user connects to the libp2p network directly instead of using registrar. Invitation link format changed. User csr is now saved to database.
 
 * Fixed android stucking on username registration screen introduced in previous alpha.
@@ -17,3 +19,9 @@
 * Labels for unregistered and duplicate usernames with modals
 
 * Fixed LoadingPanel useEffect bug.
+
+* Use csrs instead of certificates as a source of user data
+
+* Integration state manager layer with UI layer(desktop and mobile)
+
+* Clarify autoupdate language in update modal to let users know that the app will update on restart.